Integrated TinyMCE as a rich-text editor for content editing.
Replaced the old, confusing image upload process with a very slick image browser. When you click the image button on the editor you will be presented with a modal window with all images uploaded to substruct. Clicking one allows you to edit properties, then insert into content.

Will be adding an image-upload capability to the modal dialog as well.

This closes out issue 67.