Version bump to 1.3.2 for new release being worked on.
== Preference.rb

Removed all the dumb "find_by_name" / .value calls and replaced with a safe method - Preference.get_value

Added a preference to determine how long we show products in the store as "new".

== Product.rb

Changed the is_published? method to display items that have gone out of stock.
Now items must be specifically marked "discontinued" from the product page if they are not to show.

Also added an in_stock? method to possibly show items out of stock in a different method on the store. Should add this to the default view.