Added simple (as can be) HTML email for sending receipts to customers.
It makes sense now, since we allow TinyMCE to insert HTML into the "thank you for your order" snippet.

Order receipt is sent as a multi-part email, with text as before, and HTML email. The HTML snippet is stripped of HTML tags to be rendered in plain text.

Still need to come up with a good way to inline CSS styles for older mail clients, but leaving that for another day.

Closes issue 215.